redis/flight_service/requirements.txt
alembic==1.13.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3

redis==5.0.1